fix(Component): move utils into component file for now
generator was throwing error in use saying it couldn't find utils file
